Fix round-6: 22ファイル・30件のバグ修正（クラッシュ・データ破損・リソースリーク・ロジック）
Python:
- face_analyzer.py: check_image_quality の戻り値タプルバグ修正（アンパック失敗クラッシュ）
- result_auto_fetcher.py: r[0]==0 のゼロ除算を > 0 チェックで防止
- scheduler.py: task関数の run_script 戻り値を無視していたバグ修正（失敗時も後続処理が走る問題）

Java - クリティカル:
- AccuracyController: recordResult/recordV2 に @Transactional 追加（部分保存によるデータ破損防止）
- RacePredictionV2Controller: confidenceLevel を Math.min/max(0,5) でクランプ（String.repeat(-1) クラッシュ防止）
- CalendarController: java.sql.Date 型チェック追加、空raceName時のisAnalyzed誤判定修正
- EntryController: exitCode チェック追加（スクリプト失敗でも「取得完了」と表示していた問題）
- PredictionController: notifier.py ProcessBuilder に workDir 設定漏れ修正
- HealthController: dirSizeBytes でシンボリックリンクをスキップ（無限再帰防止）

Java - SseEmitter リソースリーク:
- HighDividendController/WeeklyController: onCompletion/onTimeout/onError ハンドラー追加

Java - Entity:
- HorseFaceFeature: @PrePersist 追加（createdAt が null になる問題）
- RaceSpecificPrediction: @PrePersist 追加（analyzedAt が null になりソート不定）
- RaceSpecificAccuracy: @PrePersist 追加、PK Integer→Long 統一
- RaceSpecificAccuracyRepository: 型パラメータを Long に合わせる

Java - ビジネスロジック:
- PredictionService.buildDiffStats: loser 側にのみ存在する特徴値が diffDist に含まれないバグ修正
- PredictionResultRepository: 全件取得 → findTop50ByOrderByCreatedAtDesc() に変更
- PredictionService: findLatestResults → findTop50ByOrderByCreatedAtDesc() 呼び出し

設定・その他:
- application.properties: Thymeleaf キャッシュデフォルト false→true、health details always→when_authorized
- WebMvcConfig: アップロードファイルのキャッシュ期間 3600→0（差し替え時に古い画像が返る問題）
- UploadDirInitializer: System.out/err → SLF4J ロガーに統一
- StatsPredictionController: JSON解析エラーを握りつぶさず警告ログ出力
- horse/detail.html: similar == null チェック追加（NPE防止）

+++ src/main/java/com/faceprediction/repository/RaceSpecificAccuracyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.faceprediction.entity.RaceSpecificAccuracy;
 
-public interface RaceSpecificAccuracyRepository extends JpaRepository<RaceSpecificAccuracy, Integer> {
+public interface RaceSpecificAccuracyRepository extends JpaRepository<RaceSpecificAccuracy, Long> {
 
     /** 1着的中率を出すための集計：[totalRaces, winHits] */
     @Query("SELECT COUNT(DISTINCT a.raceName), " +
